Refactor code to use method bindings for better encapsulation
The code changes in `working_with_method_bindings.py` introduce method bindings to improve encapsulation and code organization. The `function` and `function2` functions are now defined within a class called `Class`, allowing them to be accessed as methods. Additionally, a new method called `method` is added to the `Class` class.